[Platinum IV] Title: 감옥 건설, Time: 176 ms, Memory: 14044 KB -BaekjoonHub

+++ 백준/Platinum/2254. 감옥 건설/감옥 건설.java
@@ -7,7 +7,7 @@ import java.util.Stack;
 import java.util.StringTokenizer;
 
 public class Main {
-	private static final long INF = Integer.MAX_VALUE;
+	private static final long INF = Long.MAX_VALUE;
 	private static final Point BASE = new Point(INF, INF);
 	
 	private static class Point implements Comparable<Point> {
